Refactor rename CustomLogInPage and Add Link Router
